requirements.txt
raganything[all] 
fastapi 
uvicorn 
python-dotenv
pydantic
redis
sentence-transformers
transformers
qdrant-client 
psycopg2-binary
